Split module_dispatch_specs.rs into focused sub-modules (#595)
Split the monolithic module_dispatch_specs.rs (~49KB) into a directory
module with four focused sub-modules grouped by concern:

- neuron_specs.rs — 14 neuron-focused dispatch specs
- synapse_specs.rs — 7 synapse-focused dispatch specs
- structural_specs.rs — 5 structural discovery dispatch specs
- scoring_specs.rs — 10 scoring and recommendation dispatch specs

Public API remains unchanged; all existing tests pass without modification.
